Added the 1.0.0 release of MINpuT
MINpuT is a helper script for generating the structural input files used in MINT and MINT-Sandwich calculations. The output formats for the structure files include CIF, XSF, PWscf, POSCAR, and JDFTx. Users specify the flake/substrate system they wish to generate using an input file, the format of which is described in the included README.